HtcVive Update

+++ common/include/ApeIManualTexture.h
@@ -34,20 +34,20 @@ namespace Ape
 {
 	struct ManualTextureParameters
 	{
-		float height;
-		float width;
+		unsigned int height;
+		unsigned int width;
 		Ape::Texture::PixelFormat pixelFormat;
 		Ape::Texture::Usage usage;
 
 		ManualTextureParameters()
 		{
-			this->height = 0.0f;
-			this->width = 0.0f;
+			this->height = 0;
+			this->width = 0;
 			pixelFormat = Ape::Texture::PixelFormat::PF_INVALID;
 			usage = Ape::Texture::Usage::U_INVALID;
 		}
 
-		ManualTextureParameters(float height, float width, Ape::Texture::PixelFormat pixelFormat, Ape::Texture::Usage usage)
+		ManualTextureParameters(unsigned int height, unsigned int width, Ape::Texture::PixelFormat pixelFormat, Ape::Texture::Usage usage)
 		{
 			this->height = height;
 			this->width = width;
@@ -64,7 +64,7 @@ namespace Ape
 		virtual ~IManualTexture() {};
 
 	public:
-		virtual void setParameters(float width, float height, Ape::Texture::PixelFormat pixelFormat, Ape::Texture::Usage usage) = 0;
+		virtual void setParameters(unsigned int width, unsigned int height, Ape::Texture::PixelFormat pixelFormat, Ape::Texture::Usage usage) = 0;
 
 		virtual Ape::ManualTextureParameters getParameters() = 0;
